Move Negative annotation to annotations package
